* changes addToRecentAtLocalStorage callback from a custom hook to a regular JS function
